This is a 4-bit clock
The blue LEDs pins(6,9) represent minutes, the red LEDs pins(3,5) represent seconds, after each second  the second red led is blinked , and so is the case with the blue one. we can indicate odd and even minutes and seconds with this project, even use this to keep precise measure of time in minutes and seconds.